+++ ttiquete/app.json
@@ -4,7 +4,7 @@
     "slug": "ttiquete",
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "../assets/images/icon.png",
+    "icon": "./assets/images/icon.png",
     "scheme": "ttiquete",
     "userInterfaceStyle": "automatic",
     "newArchEnabled": true,
