+++ ttiquete/app/(tabs)/_layout.tsx
@@ -19,7 +19,7 @@ export default function TabLayout() {
     }}
   >
       <Tabs.Screen
-        name="home"
+        name="index"
         options={{
           title: 'Home',
           tabBarIcon: ({ color, focused }) => (
