+++ ttiquete/tsconfig.json
@@ -3,7 +3,9 @@
   "compilerOptions": {
     "strict": true,
     "paths": {
-      "@/": ["./src/"]
+      "@/*": [
+        "./*"
+      ]
     }
   },
   "include": [
